--- a/pkg/domain/domain.go
+++ b/pkg/domain/domain.go
@@ -459,6 +459,7 @@ func (do *Domain) tryLoadSchemaDiffs(m *meta.Meta, usedVersion, newVersion int64
 		if diff == nil {
 			// Empty diff means the txn of generating schema version is committed, but the txn of `runDDLJob` is not or fail.
 			// It is safe to skip the empty diff because the infoschema is new enough and consistent.
+			logutil.BgLogger().Info("diff load InfoSchema get empty schema diff", zap.Int64("version", usedVersion))
 			do.infoCache.InsertEmptySchemaVersion(usedVersion)
 			continue
 		}