@crazycs520 2024-05-23
> How about insert a clone of previous infoschema for the empty schema diff? with its schema version = previous schema version + 1 The data structure and all the other operations can be the same as before.

Nice catch, I try to use methold, but this will need to introduce another TiKV RPC to get the schemaTS of the empty schema version.